Add -git version of linux-mainline-anbox
The critical `CONFIG_ANDROID_BINDER_DEVICES` flag is set to `””` instead of the correct `”/dev/binder”` which results in failure of Anbox to recognize that it’s enabled. So I submitted a fixed -git version of this kernel to the AUR to correct this and therefore also hope to see it in Chaotic.